LLMOrchestrator/examples/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0 